+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [26.10] - 2026-10 [Unreleased]
+## [27.04] - 2027-04 [Unreleased]
+
+## [26.10] - 2026-09-20
 
 ### Added
 
@@ -658,7 +660,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - CEdit is used as base for the output
 - Visual C++ 5.0
 
-[Unreleased]: https://github.com/antonvw/wex/compare/v26.04..HEAD
+[Unreleased]: https://github.com/antonvw/wex/compare/v26.10..HEAD
+[26.10]:      https://github.com/antonvw/wex/compare/v26.04...v26.10
 [26.04]:      https://github.com/antonvw/wex/compare/v25.10...v26.04
 [25.10]:      https://github.com/antonvw/wex/compare/v25.04...v25.10
 [25.04]:      https://github.com/antonvw/wex/compare/v24.10...v25.04